Initial use of microversion_parse
microversion_parse is a new simple library that does one job: it
extracts microversions from headers. By default it will look for
the new style 'OpenStack-API-Version: compute 2.15' but can be
called to look in a variety of old style headers.

In this change 'X-OpenStack-Nova-API-Version: 2.15' continues to
work and no further changes are done. Followup patches will add
tests for the new style header and versioning of the microversion
(if required).

Depends-On: Iea8bf5cc70ad24bb352347ef347be71817db3dc5
Change-Id: I9098ab8f0800fb0550887f5eff219b18addde709