Remove api_rate_limit config option
The option was disabled by default in Havana with commit
cfdc59860ef1482e3b064f90f719ec9e484590ce and it's usage
was removed in Newton with commit
5a64f5782452d59d6ac25379a2420585e861af5a, so remove the
option since it's not longer used.

A release note is added to note the removal of the option
which was broken and confusing anyway.

Part of blueprint remove-legacy-v2-api-code

Change-Id: I23a390e5fa269c6a4a741bd414153764ea37e79c