update wording around pep8 exceptions
At least once a cycle we get contributions from new folks that attempt
to update a bunch of pep8 formatting. They believe that this is in
good faith helping, but don't realize we really aren't likely to
accept that.

Update the text here to explain that more clearly, so that people
don't keep falling down this rabbit hole.

Change-Id: Iec383a18cb3d64b9655e5bd1d82cabd3e785686c