Additional logging for placement API
This logs the allocations when we create or delete them.

This also logs some of the exceptions as LOG.exception. While they are
not always fatal or problematic, during this first cycle with the
placement-api being a thing, being more loud about odd behavior is
good. It took a lot of staring at the logs to realize how often we
were erroring in non useful ways that weren't entirely obvious.

Add __repr__ to both AllocationList and UsageList to make building
sensible log messages much simpler.

Change-Id: I8dc77408eff79b35845f6b9b5dcea6ed211cc858