[placement] functional test for report client
This sets up a scheduler report client that talks to a real
placement API (running via a wsgi-intercept). It then calls methods
to verify that a resource provider is created, some inventory set,
and some allocations written and deleted. For now this checks
simply that when things are done right, the right things happen
without any explosions. Additional tests could be added for unhappy
paths.

Some mocking of keystoneauth1 is required to get auth and endpoint
handling to work properly without a real keystone.

wsgi-intercept is added to requirements.txt to make the use of it in
these tests explicit. wsgi-intercept is already in global
requirements and is already required in nova (at a higher version)
by gabbi.

Change-Id: Ic28ace11796d7d746a16bdfa27cef9b8640f8c0e