update policy UT fixtures
RoleBasedPolicyFixture is opening a policy file specified by a conf
setting. This only works for NoDbTestCase today because of the way
TestCase sets up CONF. And now that policy has moved into code, we
really have no reason to read from a file here. We should read
defaults from the code. This makes that change.

Change-Id: I8feba7d694cc0f2971b0fb0dbe0409c90809df98