Make api_samples tests use simple cell environment
The api_samples tests depend on a very static environment where
we use the SpawnIsSynchronous fixture. This is somewhat at odds with
the locking used in the CellDatabase fixture, which will necessarily
recurse at times. Since the api_samples tests are already a bit naive
in the service layout, and since they don't care about cells things,
we do some monkeywork here to make them use the SingleCellSimple
fixture instead of a full-blown multi-cell environment.

Change-Id: If1138331f3a46f5aed87e898ce19879a787d435f