[placement] fix 500 error when allocating to bad class
Adjust exception handling when calling set_allocations so that a
KeyError in the usage_map raises an InvalidInventory. When making
allocations against a resource provider with >1 resource classes
and where one of those resource classes does not have inventory on the
provider, we can attempt to get info out of the usage_map that is not
there, and get a KeyError. This catches the KeyError and turns it into
an InvalidInventory which eventually results in a 409 response,
consistent with other responses to bad allocations. Since this is fixing
a 500, no microversion required.

Change-Id: I52fa02b56f8e62dfa206a3969a99fab250508760
Closes-Bug: #1704574