Add slowest command to tox.ini
This commit adds a printing of the 10 slowest tests during the run at
the end of the tox jobs. People have found this useful in the past and
ostestr used to do this by defualt. But, now that stestr is used to run
tests we need to make this call after a run to retain the output.

Change-Id: Ife81963e5ffb63e4ea09817e735eb768d7cf6209