Use ksa adapter for keystone conf & requests
nova.api.openstack.identity.verify_project_id now uses the common
get_ksa_adapter utility to create an Adapter from common keystoneauth1
configuration options.  This is effectively identical to the prior
functionality, with the addition of support for Adapter conf options
such as `region_name` and `valid_interfaces` which should almost never
need to be overridden.

As part of blueprint use-ksa-adapter-for-endpoints, this provides a
consistent mechanism for endpoint communication from Nova.

Change-Id: I2204c8bed8936d5bed0f410284d2a563f84e7100
Partial-Implements: bp use-ksa-adapter-for-endpoints