doc: make host aggregates examples more discoverable
When one is looking for an explanation and examples of
how to associate host aggregates to flavors, it's useful
to be able to point to that from the more top-level
aggregates page, which is linked from the main home page.

This change links from that to a somewhat buried reference
doc on host aggregates, which is actually really useful here.

Change-Id: I9e03c6649afd9a7aa816b9af1d931f9e0cffb94a