Add 'done' to migration_get_in_progress_by_host_and_node filter
This DB API method is returning all in-progress migration
records for a given host and node.

It determines what is in-progress by filtering out migration
status values for things that mean NOT in progress.

However, 'done' is an end state for a successful evacuation,
so it should be included in the list of things that are not in
progress.

Change-Id: I294fb28d279979f6e3409f1530a553dd525e53da