tests for alloc candidates with nested and traits
Adds functional tests for the AllocationCandidates.get_by_filters()
function that includes a required trait in addition to resource amounts
and the deployment has nested resource providers.

The functionality is yet to be implemented, so the test asserts
currently incorrect behaviour but is well-commented so reviewers should
be able to figure out what the scenario we're trying to isolate is.

Change-Id: Id1463f2e6fbb80eea9241dd1f6e22b6b99c4ee54
blueprint: nested-resource-providers-allocation-candidates