SchedulerReportClient._delete_provider
SchedulerReportClient._delete_provider is a new private helper method to
delete a resource provider, appropriately logging errors and raising
appropriate exceptions.

The existing public delete_resource_provider method is refactored to use
the new helper in a backward-compatible way.  In a future patch, we
should a) rename this method to delete_compute_provider or similar, and
b) make it actually raise exceptions on error (today it just logs and
carries on).

Change-Id: I05c838faca626d2ef238157a8d8cdee5df16aa01