Add functional tests for traits-based scheduling
This adds a couple of functional tests for the end-to-end
flow of creating a server with a flavor that has a required
trait. Each test has two computes. One test covers the case
of a successful placement on a compute that has a trait where
the other compute doesn't. The other test covers the case that
none of the computes have the required trait so the server
create request fails.

Part of blueprint request-traits-in-nova

Change-Id: Iacb9808ef7188e3419abfac9e8c5fb5a46c71c05