Reset the _RC_CACHE between tests
Very rarely the _RC_CACHE used for caching ResourceClass id and name
mappings will be wrong for tests, resulting in the find() method on
InventoryList returning None leading to inv.capacity calls failing.

This only showed up in the tests for
Iea182341f9419cb514a044f76864d6bec60a3683 where the order of tests
are changed because of a lot of renaming of modules.

Change-Id: Idc318d3914fa600deff613d8a43eadd9073fa262
Closes-Bug: #1747003