placement: resource requests for nested providers
Reworks the way that we solve requests for resource when nested resource
providers are present. When a granular request group is *not* sent, we
ensure that each separate resource class in the request is met by a
single provider (not summing inventories across the tree). For each
separate resource class in the request, we make a separate query to
determine the provider and root provider IDs of those resource providers
which can satisfy that request for a particular resource.

We then filter out providers that are in trees that did *not* have all
resources met by some provider in the tree. After that, we further
filter for the providers that are in trees that contain all required
traits and that the provider of those traits were the providers giving
inventory to the request.

Co-Authored-By: Balazs Gibizer <balazs.gibizer@ericsson.com>
Change-Id: I343a0cb19f4037ddde5c5fc96d0a053f699f5257
blueprint: nested-resource-providers-allocation-candidates