[placement] Fix bad management of _TRAITS_SYNCED flag
Placement functional tests could, depending on run order, be unable
to synchronise the os-traits because a previous test or this test
has failed to reset the _TRAITS_SYNCED flag.

This change fixes it so the tests will always reset the flag at
both setUp and tearDown.

In the process a nearby misleading comment which says a database is
not being used is corrected.

Change-Id: I595be7bca2c1bde86651b126ce501286b301d272
Closes-Bug: #1759863