update scheduler to use image-traits
Update the scheduler to take into account the traits specified
as part of the image properties. The required traits from the
image are added to the resource request along with the traits
and resources from the flavor.

This change does not handle rebuilding of the instance, this
will come as part of a follow on patch.

Added unit and functional tests

Change-Id: I3e00080a6b557c6a27634b1dc04fc589f0da0bc2
Implements: blueprint glance-image-traits