Remove support for /os-fping REST API
This drops support for the fping compute REST API which
has been deprecated since Newton:

  I1a8a44530be29292561e90d6f7bd7ed512a88ee3

To match the os-cloudpipe and os-certificates removals
the os-fping controller now returns a 410 response.

The related fping_path configuration option is removed
along with the related fping policy rules.

Unit tests are removed and the functional API sample
test is just asserting the 410 response now.

The API sample docs are left intact since the API reference
still builds from those and can be considered more or
less branchless, so people looking at the API reference
can apply it to older deployments of nova before os-fping
was removed.

A release note is started which we can build on for each
nova-network specific API that we remove in this series.

Part of blueprint remova-nova-network

Change-Id: Ia36aaa8f74adc2b540c49523db522cd85ab17ed2