Extract part of PlacementFixture to placement
In nova.tests.fixtures is a PlacementFixture that peforms two roles. One
is to provide a working placement service via wsgi-intercept. The other
is to provide a working client of that service in the form of a
scheduler report client that uses appropriate headers and endpoints.

This change extracts the first role to a new fixture hosted within the
placement hierarchy and makes the second role a subclass of the first.

To make this work nicely, existing placement fixtures (in fixtures.py),
used solely for the gabbi tests are moved into a new directory called
"fixtures" and renamed to "gabbits.py". "gabbi.py" can't be used
because of naming conflicts while importing.

This is a small part of ongoing work related to
blueprint placement-extract

Change-Id: I126ada549d3879f89d1ec64b743da14807a39351